requirements.txt
librosa
soundfile
numpy
scipy
matplotlib
mutagen
Pillow